test(auth): split status and integration auth tests below 4k (#1399)
